Final abstractions for BaseSwerveModule/Subsystem
